extract attribute schema
